backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
Django==4.1.3
django-extensions==3.2.1
django-jazzmin==2.6.0
django-nose==1.4.7
djangorestframework==3.14.0
idna==3.4
nose==1.3.7
psycopg2==2.9.5
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
